use a callback for when a module ends
